--- dist.bat
@@ -1,5 +1,0 @@
-
-cp -R build/libs/. c:/users/lothr/desktop
-
-rm -Rf build/libs/
-
