+++ gradle.properties
@@ -3,12 +3,17 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
+# Base info
 curse_id=233659
 
+mc_version=1.16.5
+forge_version=36.2.0
+mcp_channel=snapshot
+mcp_mappings=20210309-1.16.5
 
-mc_version=1.14.4
+# Version
 mod_version=1.0.22
-forge_version=28.1.111
 
-jei_version=6.0.0.11
+# Dependencies
+jei_version=7.7.0.107
