--- langCopy.bat
@@ -1,4 +1,0 @@
-cp src/main/resources/assets/similsaxtranstructors/lang/en_AU.lang src/main/resources/assets/similsaxtranstructors/lang/en_CA.lang
-cp src/main/resources/assets/similsaxtranstructors/lang/en_AU.lang src/main/resources/assets/similsaxtranstructors/lang/en_GB.lang
-cp src/main/resources/assets/similsaxtranstructors/lang/en_AU.lang src/main/resources/assets/similsaxtranstructors/lang/en_US.lang
-cp src/main/resources/assets/similsaxtranstructors/lang/en_AU.lang src/main/resources/assets/similsaxtranstructors/lang/en_NZ.lang
