+++ src/main/java/com/vorquel/similsaxtranstructors/SimilsaxTranstructors.java
@@ -1,5 +1,8 @@
 package com.vorquel.similsaxtranstructors;
 
+import com.vorquel.similsaxtranstructors.client.BlockOverlay;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.fml.DistExecutor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import net.minecraft.item.Item;
@@ -14,27 +17,30 @@ import net.minecraftforge.fml.loading.FMLPaths;
 @Mod(SimilsaxTranstructors.MODID)
 public class SimilsaxTranstructors {
 
-  public static final String MODID = "similsaxtranstructors";
-  static Logger log = LogManager.getLogger(MODID);
+	public static final String MODID = "similsaxtranstructors";
+	static Logger log = LogManager.getLogger(MODID);
 
-  public SimilsaxTranstructors() {
-    FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
-    MinecraftForge.EVENT_BUS.register(this);
-    MinecraftForge.EVENT_BUS.register(new BlockOverlay());
-    ConfigHandler.loadConfig(ConfigHandler.COMMON_CONFIG, FMLPaths.CONFIGDIR.get().resolve(MODID + ".toml"));
-  }
+	public SimilsaxTranstructors() {
+		FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
+		MinecraftForge.EVENT_BUS.register(this);
+		ConfigHandler.loadConfig(ConfigHandler.COMMON_CONFIG, FMLPaths.CONFIGDIR.get().resolve(MODID + ".toml"));
 
-  private void setup(final FMLCommonSetupEvent event) {}
+		DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> {
+			MinecraftForge.EVENT_BUS.register(new BlockOverlay());
+		});
+	}
 
-  @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
-  public static class RegistryEvents {
+	private void setup(final FMLCommonSetupEvent event) {}
 
-    @SubscribeEvent
-    public static void onItemsRegistry(final RegistryEvent.Register<Item> event) {
-      event.getRegistry().register(new ItemSimilsax(new Item.Properties().maxStackSize(1)
-          .maxDamage(800)).setRegistryName("transtructor_basic"));
-      event.getRegistry().register(new ItemSimilsax(new Item.Properties().maxStackSize(1)
-          .maxDamage(9000)).setRegistryName("transtructor_advanced"));
-    }
-  }
+	@Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
+	public static class RegistryEvents {
+
+		@SubscribeEvent
+		public static void onItemsRegistry(final RegistryEvent.Register<Item> event) {
+			event.getRegistry().register(new ItemSimilsax(new Item.Properties().maxStackSize(1)
+					.maxDamage(800)).setRegistryName("transtructor_basic"));
+			event.getRegistry().register(new ItemSimilsax(new Item.Properties().maxStackSize(1)
+					.maxDamage(9000)).setRegistryName("transtructor_advanced"));
+		}
+	}
 }
