--- src/main/java/com/vorquel/similsaxtranstructors/setup/IProxy.java
@@ -1,10 +1,0 @@
-package com.vorquel.similsaxtranstructors.setup;
-
-import net.minecraft.world.World;
-
-public interface IProxy {
-	World getClientWorld();
-
-
-	public void registerBlockOverlay();
-}
