--- src/main/java/com/vorquel/similsaxtranstructors/setup/ServerProxy.java
@@ -1,16 +1,0 @@
-package com.vorquel.similsaxtranstructors.setup;
-
-import net.minecraft.world.World;
-
-public class ServerProxy implements IProxy {
-	@Override
-	public World getClientWorld() {
-		 throw new IllegalStateException("Client side only code on the server");
-	}
-
-
-
-	public void registerBlockOverlay(){
-
-	}
-}
