--- src/main/resources/mcmod.info
@@ -1,21 +1,0 @@
-[
-{
-  "modid": "similsaxtranstructors",
-  "name": "Similsax Transtructors",
-  "description": "Now you can place blocks *behind* the block you click on!",
-  "version": "1.12-1.0.21",
-  "mcversion": "1.12",
-  "url": "",
-  "updateUrl": "",
-  "authorList": [
-    "Vorquel", "Lothrazar"
-  ],
-  "credits": "Thanks to all my fans and sponsors.",
-  "logoFile": "",
-  "screenshots": [
-  ],
-  "parent": "",
-  "dependencies": [
-  ]
-}
-]
