+++ src/main/resources/pack.mcmeta
@@ -1,7 +1,6 @@
 {
     "pack": {
-        "description": "similsaxtranstructors resources",
-        "pack_format": 4,
-        "_comment": "A pack_format of 4 requires json lang files. Note: we require v4 pack meta for all mods."
+        "description": "Similsax Transtructors resources",
+        "pack_format": 6
     }
 }
